requirements.txt
torch
torchvision
scipy
matplotlib
Pillow
scikit-image
opencv-python
